requirement.txt
openai
python-dotenv
langchain_community
langchain_openai
chromadb == 0.3.29